+++ CHANGELOG.md
@@ -8,6 +8,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Remove support for Helm 2
 - Upgrade to go 1.18.1
 - Upgrade go-yaml.v2 to 2.4.0
+- Fix build to properly include VERSION
 
 ## [0.7.1] - 2022-05-05